backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
httpx>=0.27
psycopg2-binary>=2.9
python-dotenv>=1.0
apscheduler>=3.10
joblib>=1.3
xgboost>=2.0
scikit-learn>=1.3
numpy>=1.24
pandas>=2.0
shap>=0.44
pytz>=2024.1
